# openapi.json
{"openapi":"3.1.0","info":{"title":"beancounters","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/accounts":{"get":{"tags":["api"],"operationId":"list_accounts","responses":{"200":{"description":"List all accounts","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}},"500":{"description":"Internal server error"}}},"post":{"tags":["api"],"operationId":"add_account","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"201":{"description":"Account created"},"500":{"description":"Internal server error"}}}},"/accounts/{name}":{"put":{"tags":["api"],"operationId":"update_account","parameters":[{"name":"name","in":"path","description":"Account name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"Account updated"},"500":{"description":"Internal server error"}}},"delete":{"tags":["api"],"operationId":"delete_account","parameters":[{"name":"name","in":"path","description":"Account name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Account deleted"},"500":{"description":"Internal server error"}}}},"/accounts/{name}/close":{"post":{"tags":["api"],"operationId":"close_account","parameters":[{"name":"name","in":"path","description":"Account name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseAccountRequest"}}},"required":true},"responses":{"200":{"description":"Account closed"},"500":{"description":"Internal server error"}}}},"/transactions":{"get":{"tags":["api"],"operationId":"list_transactions","responses":{"200":{"description":"List all transactions","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"500":{"description":"Internal server error"}}},"post":{"tags":["api"],"operationId":"add_transaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"201":{"description":"Transaction created"},"500":{"description":"Internal server error"}}}},"/transactions/{id}":{"put":{"tags":["api"],"operationId":"update_transaction","parameters":[{"name":"id","in":"path","description":"Transaction ID","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"200":{"description":"Transaction updated"},"500":{"description":"Internal server error"}}},"delete":{"tags":["api"],"operationId":"delete_transaction","parameters":[{"name":"id","in":"path","description":"Transaction ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction deleted"},"500":{"description":"Internal server error"}}}},"/transactions/{id}/clear":{"post":{"tags":["api"],"operationId":"clear_transaction","parameters":[{"name":"id","in":"path","description":"Transaction ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction cleared"},"500":{"description":"Internal server error"}}}},"/transactions/{id}/unclear":{"post":{"tags":["api"],"operationId":"unclear_transaction","parameters":[{"name":"id","in":"path","description":"Transaction ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction uncleared"},"500":{"description":"Internal server error"}}}},"/verify":{"get":{"tags":["api"],"operationId":"verify_ledger","responses":{"200":{"description":"Verify ledger","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResult"}}}},"500":{"description":"Internal server error"}}}}},"components":{"schemas":{"Account":{"type":"object","required":["name","open_date","currencies"],"properties":{"close_date":{"type":["string","null"]},"currencies":{"type":"array","items":{"type":"string"}},"name":{"type":"string"},"open_date":{"type":"string"}}},"CloseAccountRequest":{"type":"object","required":["date"],"properties":{"date":{"type":"string"}}},"Posting":{"type":"object","required":["account","amount","currency"],"properties":{"account":{"type":"string"},"amount":{"type":"string"},"cost":{"type":["string","null"]},"currency":{"type":"string"},"price":{"type":["string","null"]}}},"Transaction":{"type":"object","required":["date","flag","tags","postings"],"properties":{"date":{"type":"string"},"flag":{"type":"string"},"id":{"type":["string","null"]},"narration":{"type":["string","null"]},"payee":{"type":["string","null"]},"postings":{"type":"array","items":{"$ref":"#/components/schemas/Posting"}},"tags":{"type":"array","items":{"type":"string"}}}},"VerifyResult":{"type":"object","required":["errors","warnings"],"properties":{"errors":{"type":"array","items":{"type":"string"}},"warnings":{"type":"array","items":{"type":"string"}}}}}},"tags":[{"name":"beancounters","description":"Beancount API"}]}